+++ src/ExportEdiJobs/ExportEdiJobs.test.js
@@ -7,6 +7,11 @@ import { runAxeTest } from '@folio/stripes-testing';
 import { useExportEdiJobsQuery } from './apiQuery';
 import { ExportEdiJobs } from './ExportEdiJobs';
 
+jest.mock('@folio/stripes/core', () => ({
+  ...jest.requireActual('@folio/stripes/core'),
+  TitleManager: ({ children }) => <>{children}</>,
+}), { virtual: true });
+
 jest.mock('react-virtualized-auto-sizer', () => {
   return jest.fn(({ children }) => <div>{children({})}</div>);
 });