+++ src/common/constants.js
@@ -15,7 +15,14 @@ export const EXPORT_JOB_STATUS_OPTIONS = Object.values(EXPORT_JOB_STATUSES).map(
 
 export const EXPORT_CONFIGS_API = 'data-export-spring/configs';
 export const EXPORT_JOBS_API = 'data-export-spring/jobs';
+export const MULTIPLE_EXPORTED_JOB_TYPES = [
+  'BULK_EDIT_IDENTIFIERS',
+  'BULK_EDIT_QUERY',
+  'BULK_EDIT_UPDATE',
+  'BURSAR_FEES_FINES',
+];
 export const EXPORTED_JOB_TYPES = [
+  ...MULTIPLE_EXPORTED_JOB_TYPES,
   'CIRCULATION_LOG',
   'E_HOLDINGS',
   'AUTH_HEADINGS_UPDATES',