+++ src/common/hooks/useSecureDownload/useSecureDownload.js
@@ -13,13 +13,15 @@ export const useSecureDownload = (jobId) => {
   const ky = useOkapiKy();
   const showCallout = useShowCallout();
 
-  const download = async (fileName) => {
-    return ky.get(`${EXPORT_JOBS_API}/${jobId}/download`, {
+  const downloadSingleFile = async (fileName, parameterized) => {
+    const key = parameterized ? `key=${encodeURIComponent(fileName)}` : '';
+
+    return ky.get(`${EXPORT_JOBS_API}/${jobId}/download?${key}`, {
       headers: { accept: 'application/octet-stream' },
     })
       .blob()
       .then(data => {
-        downloadBase64(fileName, URL.createObjectURL(data));
+        downloadBase64(fileName.replace(`${jobId}/`, ''), URL.createObjectURL(data));
       })
       .catch(() => {
         showCallout({
@@ -28,6 +30,10 @@ export const useSecureDownload = (jobId) => {
         });
       });
   };
+
+  const download = async (fileNames, parameterized) => {
+    return Promise.all(fileNames.map((fileName) => downloadSingleFile(fileName, parameterized)));
+  }
 
   return { download };
 };